a
Tried adding a coral Subsystem, It kind of worked. Although coral system is probably a better name. I'm stiil not sure why the internal run command doesn't get called at all, but it's like 2 and I need to be up at 9 for the meeting.
